add & render toilet
